+++ app/src/constants.rs
@@ -1,2 +1,5 @@
-pub const SNAPSHOT_FILENAME: &str = "snapshot";
+pub const DEFAULT_CONFIG_FILE: &str = "config.ini";
+pub const DEFAULT_BINARY_DIR: &str = ".";
+pub const DEFAULT_SNAPSHOT_TAG: &str = "snapshot";
+
 pub const SNAPSHOT_EXTENSION: &str = "json";