+++ app/src/lib.rs
@@ -51,17 +51,32 @@ pub fn execute(options: &Options) -> Result<(), String> {
                 users::LinuxUsers,
                 root::LinuxRoot,
                 tcp_listen::LinuxTCPListen,
+                suid::LinuxSuid,
+                guid::LinuxGuid,
+                promisc::LinuxPromisc,
+                crontab::LinuxCrontab,
+                dns::LinuxDns,
             };
 
             // Instantiate Linux plugins
             let linux_users = LinuxUsers::new();
             let linux_root = LinuxRoot::new();
             let linux_tcp_listen = LinuxTCPListen::new();
+            let linux_suid = LinuxSuid::new();
+            let linux_guid = LinuxGuid::new();
+            let linux_promisc = LinuxPromisc::new();
+            let linux_crontab = LinuxCrontab::new();
+            let linux_dns = LinuxDns::new();
             let plugins: Vec<&dyn Plugin> = vec![
                                                     // Execute Linux plugins
                                                     &linux_users,
-                                                    &linux_tcp_listen,
                                                     &linux_root,
+                                                    &linux_tcp_listen,
+                                                    &linux_suid,
+                                                    &linux_guid,
+                                                    &linux_promisc,
+                                                    &linux_crontab,
+                                                    &linux_dns,
                                                 ];
         } else if #[cfg(target_os = "macos")] {
             use crate::plugins::macos::{
@@ -83,11 +98,13 @@ pub fn execute(options: &Options) -> Result<(), String> {
         Mode::Run => run(
             &plugins, 
             &options.config, 
-            &options.binary_directory, 
+            &options.binary_directory,
+            &options.snapshot_tag,
             &options.verbose
         ),
         Mode::Merge => merge(
-            &options.merging_directory, 
+            &options.merging_directory,
+            &options.snapshot_tag,
             &options.verbose
         ),
         Mode::Compare => compare(