+++ app/src/plugins/linux/mod.rs
@@ -1,3 +1,8 @@
-pub mod users;
-pub mod tcp_listen;
+pub mod crontab;
+pub mod dns;
+pub mod guid;
+pub mod promisc;
 pub mod root;
+pub mod suid;
+pub mod tcp_listen;
+pub mod users;