+++ rusthunter.ps1
@@ -25,7 +25,10 @@ Param(
 
     [string]
     $ConfigFile = $null,
-    
+
+    [string]
+    $SnapshotTag = $null,
+
     [string]
     $InitialSnapshot = $null,
 
@@ -69,7 +72,7 @@ ${MACOS_BINARIES_PATH}="${ANSIBLE_PATH}\roles\macos\files"
 ${WINDOWS_BINARIES_PATH}="${ANSIBLE_PATH}\roles\windows\files"
 ${SNAPSHOT_PATH}=".\launcher\snapshots"
 
-${DEFAULT_CONFIG_FILE}="config"
+${DEFAULT_CONFIG_FILE}="config.ini"
 ${DEFAULT_HOSTS_FILE}="hosts"
 ###########################################################
 
@@ -114,11 +117,13 @@ function Show-Help {
     Write-Host "usage: $0 local -ConfigFile CONFIG_FILE"
     Write-Host
     Write-Host "     -ConfigFile         Configuration file"
+    Write-Host "     -SnapshotTag        Snapshot tag"
     Write-Host
     Write-Host "usage: $0 global -HostsFile HOSTS_FILE -ConfigFile CONFIG_FILE"
     Write-Host
     Write-Host "     -HostsFile          Hosts file"
     Write-Host "     -ConfigFile         Configuration file"
+    Write-Host "     -SnapshotTag        Snapshot tag"
     Write-Host
     Write-Host "usage: $0 compare -InitialSnapshot INITIAL_SNAPSHOT -CurrentSnapshot CURRENT_SNAPSHOT -ShowStatistics -Host HOST -Plugin PLUGIN"
     Write-Host
@@ -134,11 +139,11 @@ function Show-Help {
 function Show-Error($message) {
     Write-Host " [-] ${message}" -ForegroundColor red
     Write-Host ""
-    Exit 1    
+    Exit 1
 }
 
 function Show-Warning($message) {
-    Write-Host " [!] ${message}" -ForegroundColor yellow      
+    Write-Host " [!] ${message}" -ForegroundColor yellow
 }
 
 function Show-Info($message) {
@@ -147,19 +152,19 @@ function Show-Info($message) {
 
 function Is-ExecutableInstalled {
     if ( !(Test-Path ${INSTALLATION_PATH}\${EXECUTABLE_NAME}) ) {
-        Show-Error "The tool has not been installed yet"       
+        Show-Error "The tool has not been installed yet"
     }
 }
 
 function Is-DockerInstalled {
     if ( !$(docker --version 2> $null) ){
-        Show-Error "Please install Docker Desktop for Windows"       
+        Show-Error "Please install Docker Desktop for Windows"
     }
 }
 
 function Build-BuilderImage {
     if ( !$(docker images -q ${BUILDER_IMAGE_NAME}:latest 2> $null) ) {
-        Show-Info "Building builder docker image" 
+        Show-Info "Building builder docker image"
         docker build -t ${BUILDER_IMAGE_NAME} ${BUILDER_IMAGE_PATH}
     }
 }
@@ -173,7 +178,7 @@ function Build-LauncherImage {
 
 function Install-RustHunter {
     if ( !(Test-Path ${WINDOWS_BINARIES_PATH}\${EXECUTABLE_NAME}) ){
-        Show-Error "The tool has not been built yet"       
+        Show-Error "The tool has not been built yet"
     } else {
         Show-Info "Installing executable"
         cp ${WINDOWS_BINARIES_PATH}\${EXECUTABLE_NAME} ${INSTALLATION_PATH}
@@ -186,7 +191,7 @@ function Install-RustHunter {
 
 function Show-Plugins {
     Is-ExecutableInstalled
-    
+
     Invoke-Expression "${EXECUTABLE_NAME} list"
 }
 
@@ -194,21 +199,15 @@ function Get-LocalSnapshot {
     Is-ExecutableInstalled
 
     if (!${ConfigFile}) {
-        Show-Error "Please specify a config file"
+      Show-Error "Please specify a config file"
     }
 
-    Show-Info "Creating snapshots directory"
-    mkdir -p ${SNAPSHOT_PATH} > $null
-
     Show-Info "Collecting data"
-    Invoke-Expression "${EXECUTABLE_NAME} run -c ${ConfigFile} -b ${WINDOWS_BINARIES_PATH}"
-    mv snapshot.json ${SNAPSHOT_PATH}
-
-    Show-Info "Merging data"
-    Invoke-Expression "${EXECUTABLE_NAME} merge -d ${SNAPSHOT_PATH}"
-
-    Show-Info "Cleaning up"
-    Remove-Item -Force -Recurse ${SNAPSHOT_PATH}
+    if (!${SnapshotTag}) {
+      Invoke-Expression "${EXECUTABLE_NAME} run -c ${ConfigFile} -b ${WINDOWS_BINARIES_PATH}"
+    } else {
+      Invoke-Expression "${EXECUTABLE_NAME} run -c ${ConfigFile} -b ${WINDOWS_BINARIES_PATH} --tag ${SnapshotTag}"
+    }
 }
 
 function Is-FileEncrypted($file) {
@@ -243,24 +242,24 @@ function Protect-Hosts {
         Show-Error "${HostsFile} is not encrypted"
     }
 
-    if ( ${EncryptHosts} -and ${isFileEncrypted} ) {
-        $command = "encrypt"
+    if ( ${EncryptHosts} -and !${isFileEncrypted} ) {
+        ${Command} = "encrypt"
         Show-Info "Encrypting hosts file"
     } elseif ( ${RekeyHosts} ) {
-        $command = "rekey"
+        ${Command} = "rekey"
         Show-Info "Rekeying hosts file"
     } elseif ( ${ViewHosts} ) {
-        $command = "view"
+        ${Command} = "view"
         Show-Info "Showing hosts file"
     } elseif ( ${EditHosts} ) {
-        $command = "edit"
+        ${Command} = "edit"
         Show-Info "Editing hosts file"
     } elseif ( ${DecryptHosts} ) {
-        $command = "decrypt"
+        ${Command} = "decrypt"
         Show-Info "Decrypting hosts file"
     }
 
-    docker run --rm -v $PWD\${HostsFile}:/tmp/hosts -it ${LAUNCHER_IMAGE_NAME}:latest bash -c "cp /tmp/hosts /tmp/host.tmp;env EDITOR=nano ansible-vault $command /tmp/host.tmp; cp /tmp/host.tmp /tmp/hosts"
+    docker run --rm -v $PWD\${HostsFile}:/tmp/hosts -it ${LAUNCHER_IMAGE_NAME}:latest bash -c "cp /tmp/hosts /tmp/host.tmp;env EDITOR=nano ansible-vault ${Command} /tmp/host.tmp; cp /tmp/host.tmp /tmp/hosts"
 }
 
 function Get-GlobalSnapshot {
@@ -284,17 +283,28 @@ function Get-GlobalSnapshot {
     cp ${ConfigFile} ${WINDOWS_BINARIES_PATH}/${DEFAULT_CONFIG_FILE}
 
     Show-Info "Creating snapshots directory"
-    mkdir -p ${SNAPSHOT_PATH} > $null
+    mkdir -Force -Path ${SNAPSHOT_PATH} > $null
 
     Show-Info "Collecting data"
+    ${Arguments} = ""
     if ( Is-FileEncrypted ${HostsFile} ) {
-        docker run --rm -v $PWD\${ANSIBLE_PATH}:/etc/ansible -v $PWD\${SNAPSHOT_PATH}:/snapshots -w /etc/ansible -it ${LAUNCHER_IMAGE_NAME}:latest ansible-playbook --ask-vault-pass playbook.yml
-    } else {
-        docker run --rm -v $PWD\${ANSIBLE_PATH}:/etc/ansible -v $PWD\${SNAPSHOT_PATH}:/snapshots -w /etc/ansible -it ${LAUNCHER_IMAGE_NAME}:latest ansible-playbook playbook.yml
+      ${Arguments} += " --ask-vault-pass"
     }
 
+    if (${SnapshotTag}) {
+      ${Arguments} += " --extra-vars `"snapshot_tag=${SnapshotTag}`""
+    }
+
+    Invoke-Expression "docker run --rm -v $PWD\${ANSIBLE_PATH}:/etc/ansible -v $PWD\${SNAPSHOT_PATH}:/snapshots -w /etc/ansible -it ${LAUNCHER_IMAGE_NAME}:latest ansible-playbook ${Arguments} playbook.yml"
+
     Show-Info "Merging data"
-    Invoke-Expression "${EXECUTABLE_NAME} merge -d ${SNAPSHOT_PATH}"
+    ${Arguments} = "-d ${SNAPSHOT_PATH}"
+
+    if ( ${SnapshotTag}) {
+      Invoke-Expression "${EXECUTABLE_NAME} merge -d ${SNAPSHOT_PATH} --tag ${SnapshotTag}"
+    } else {
+      Invoke-Expression "${EXECUTABLE_NAME} merge -d ${SNAPSHOT_PATH}"
+    }
 
     Show-Info "Cleaning up"
     Remove-Item -Force -Recurse ${SNAPSHOT_PATH}
@@ -303,18 +313,18 @@ function Get-GlobalSnapshot {
 function Compare-Snapshots {
     Is-ExecutableInstalled
 
-    $args = ""
+    ${Arguments} = ""
 
     if ( !${InitialSnapshot} ) {
         Show-Error "Please specify an initial snapshot file"
     } else {
-        $args += " --initial ${InitialSnapshot}"
+        ${Arguments} += " --initial ${InitialSnapshot}"
     }
 
     if ( !${CurrentSnapshot} ) {
         Show-Error "Please specify a current snapshot file"
     } else {
-        $args += " --current ${CurrentSnapshot}"
+        ${Arguments} += " --current ${CurrentSnapshot}"
     }
 
     if ( ${FilterPlugin} -and !${FilterHost}) {
@@ -322,18 +332,18 @@ function Compare-Snapshots {
     }
 
     if ( ${FilterHost} ) {
-        $args += " --host ${FilterHost}"
+        ${Arguments} += " --host ${FilterHost}"
     }
 
     if ( ${FilterPlugin} ) {
-        $args += " --plugin ${FilterPlugin}"
+        ${Arguments} += " --plugin ${FilterPlugin}"
     }
 
     if ( ${ShowStatistics} ) {
-        $args += " --stats"
+        ${Arguments} += " --stats"
     }
 
-    Invoke-Expression "${EXECUTABLE_NAME} compare $args"
+    Invoke-Expression "${EXECUTABLE_NAME} compare ${Arguments}"
 }
 
 function Uninstall-RustHunter {
@@ -375,13 +385,12 @@ function Build-RustHunter {
 }
 
 function Update-RustHunter {
+    # Requires git
     Show-Info "Downloading latest updates"
     git pull
 
     Show-Info "Installing new executable"
     cp ${WINDOWS_BINARIES_PATH}\${EXECUTABLE_NAME} ${INSTALLATION_PATH}
-    
-    Build-LauncherImage
 
     Show-Info "Successfully updated"
 }
@@ -391,8 +400,6 @@ function Test-RustHunter {
 
     Is-DockerInstalled
 
-
-
     if ( !${UnitTests} -and !${IntegrationTests} -and !${ValidationTests} ) {
         Show-Error "No tests specified"
     }
@@ -400,42 +407,40 @@ function Test-RustHunter {
     if ( ${UnitTests} ) {
         Build-BuilderImage
 
-        Show-Info "Unit testing for Linux target"
-        docker run --rm -v $PWD\${APP_PATH}:/app -w /app ${BUILDER_IMAGE_NAME}:latest cargo test --lib --target x86_64-unknown-linux-gnu
-
-        Show-Info "Unit testing for Windows target"
-        docker run --rm -v $PWD\${APP_PATH}:/app -w /app ${BUILDER_IMAGE_NAME}:latest cargo test --lib --target x86_64-pc-windows-msvc
+        Show-Info "Unit testing" # OS-independent
+        docker run --rm -v $PWD\${APP_PATH}:/app -w /app ${BUILDER_IMAGE_NAME}:latest cargo test --lib
     }
-    
+
     if ( ${IntegrationTests} ) {
         Build-BuilderImage
-        
-        Show-Info "Integration testing for Linux target"
+
+        Show-Info "Integration testing" # OS-dependent
         docker run --rm -v $PWD\${APP_PATH}:/app -w /app ${BUILDER_IMAGE_NAME}:latest cargo test --test integration
     }
 
     if ( $ValidationTests ) {
-        Build-LauncherImage
-
-        Show-Info "Creating snapshots directory"
-        mkdir -p ${SNAPSHOT_PATH} > $null
-
         Show-Info "Creating target dockers"
-        docker network create rusthunter_test_net --driver=bridge --subnet="192.168.100.1/24"
-        for ($i = 2 ; $i -le 20 ; $i++) {
-            docker run --network=rusthunter_test_net --ip="192.168.100.$i" -d peco602/ssh-linux-docker:latest
+        $N=4
+        echo "[linux]" | Out-File -Encoding ASCII test.hosts
+        for ($i = 2 ; $i -le $N ; $i++) {
+            ${TARGET_NAME}="target-$i"
+            docker run --name ${TARGET_NAME} -d peco602/ssh-linux-docker:latest
+            ${TARGET_IP}=$(docker inspect -f "{{ .NetworkSettings.Networks.bridge.IPAddress }}" $TARGET_NAME)
+            echo "${TARGET_IP} ansible_connection=ssh ansible_user=user ansible_ssh_password=Pa`$`$w0rd123! ansible_become_password=Pa`$`$w0rd123!" | Out-File -Encoding ASCII -Append test.hosts
         }
 
-        Show-Info "Collecting data"
-        docker run --rm -v $PWD\${ANSIBLE_PATH}:/etc/ansible -v $PWD\${SNAPSHOT_PATH}:/snapshots -w /etc/ansible --network=rusthunter_test_net ${LAUNCHER_IMAGE_NAME}:latest ansible-playbook playbook.yml -i hosts.test
+        ${ConfigFile} = ${DEFAULT_CONFIG_FILE}
+        ${HostsFile} = "test.hosts"
+        ${SnapshotTag} = "validation"
 
-        Show-Info "Merging data"
-        Invoke-Expression "${EXECUTABLE_NAME} merge -d ${SNAPSHOT_PATH}"
+        Get-GlobalSnapshot
 
-        Show-Info "Cleaning up"
-        docker rm $(docker network inspect rusthunter_test_net --format='{{range $id, $_ := .Containers}}{{println $id}}{{end}}') --force
-        docker network rm rusthunter_test_net
-        Remove-Item -Force -Recurse ${SNAPSHOT_PATH}
+        Show-Info "Destroying target dockers"
+        for ($i = 2 ; $i -le $N ; $i++) {
+            ${TARGET_NAME}="target-$i"
+            docker rm $TARGET_NAME --force
+        }
+        Remove-Item ${HostsFile}
     }
 }
 
@@ -458,6 +463,6 @@ switch ($Subcommand) {
     "uninstall" { Uninstall-RustHunter }
     "build"     { Build-RustHunter }
     "update"    { Update-RustHunter }
-    "test"      { Test-RustHunter } 
+    "test"      { Test-RustHunter }
     default     { Show-Help }
 }